+++ glean-core/ios/Glean/Config/Configuration.swift
@@ -8,6 +8,7 @@ public struct Configuration {
     let serverEndpoint: String
     let maxEvents: Int32?
     let channel: String?
+    let dataPath: String?
 
     struct Constants {
         static let defaultTelemetryEndpoint = "https://incoming.telemetry.mozilla.org"
@@ -18,14 +19,19 @@ public struct Configuration {
     /// - parameters:
     ///   * maxEvents the number of events to store before the events ping is sent.
     ///   * channel the release channel the application is on, if known.
-    ///     This will be sent along with all the pings, in the `client_info` section.
+    ///   This will be sent along with all the pings, in the `client_info` section.
+    ///   * serverEndpoint the server endpoint Glean should send data to
+    ///   * dataPath an optional String that specifies where to store data locally on the device.
+    ///   This should ONLY be used when setting up Glean on a non-main process.
     public init(
         maxEvents: Int32? = nil,
         channel: String? = nil,
-        serverEndpoint: String? = nil
+        serverEndpoint: String? = nil,
+        dataPath: String? = nil
     ) {
         self.serverEndpoint = serverEndpoint ?? Constants.defaultTelemetryEndpoint
         self.maxEvents = maxEvents
         self.channel = channel
+        self.dataPath = dataPath
     }
 }