+++ CHANGELOG.md
@@ -1,15 +1,28 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v52.4.3...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v52.5.0...main)
+
+# v52.5.0 (2023-04-11)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v52.4.3...v52.5.0)
+
+* General
+  * On Rkv detecting a corrupted database delete the old RKV, create a new one and log the error ([#2425](https://github.com/mozilla/glean/pull/2425))
+  * Add the Date header as late as possible before the uploader acts ([#2436](https://github.com/mozilla/glean/pull/2436))
+  * The logic of the Server Knobs API has been flipped. Instead of applying a list of metrics and their _disabled_ state, the API now accepts a list of metrics and their _enabled_ state ([bug 1811253](https://bugzilla.mozilla.org/show_bug.cgi?id=1811253))
+* Kotlin
+  * Adds the ability to record metrics on a non-main process. This is enabled by setting a `dataPath` in the Glean configuration ([bug 1815233](https://bugzilla.mozilla.org/show_bug.cgi?id=1815233))
+* iOS
+  * Adds the ability to record metrics on a non-main process. This is enabled by setting a `dataPath` in the Glean configuration ([bug 1815233](https://bugzilla.mozilla.org/show_bug.cgi?id=1815233))
 
 # v52.4.3 (2023-03-24)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v52.4.2...v52.4.3)
 
 * General
   * Expose Server Knobs functionality via UniFFI for use on mobile
 * iOS
   * BUGFIX: Prevent another test-only issue: The storage going away when the uploader reports back its status ([#2430](https://github.com/mozilla/glean/pull/2430))
-
-[Full changelog](https://github.com/mozilla/glean/compare/v52.4.2...v52.4.3)
 
 # v52.4.2 (2023-03-15)